feat: :sparkles: fetched video section from cms db
