+++ package.json
@@ -32,13 +32,13 @@
     "@radix-ui/react-separator": "^1.1.8",
     "@radix-ui/react-slot": "^1.2.4",
     "@radix-ui/react-tabs": "^1.1.13",
-    "@react-email/components": "^1.0.3",
-    "@react-email/render": "^2.0.1",
+    "@react-email/components": "^1.0.4",
+    "@react-email/render": "^2.0.2",
     "cheerio": "^1.1.2",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "crisp-sdk-web": "^1.0.27",
-    "framer-motion": "^12.24.7",
+    "framer-motion": "^12.24.10",
     "graphql": "^16.12.0",
     "lucide-react": "^0.562.0",
     "next": "16.1.1",
